requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
rich>=13.0.0
openai>=1.0.0
tomli>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
pypdf>=4.0.0
